Extract animation to own component and make it work again
